Add Python documentation outside of stub files as well
Language servers seem to prefer to show doc strings from stub files if
present while e.g., an interactive session with the Python interpreter
would ignore them an only look at what is embeded in the library.

This patch adds doc strings to the generated library as well which is a
lot of duplication, but I am not sure there is a way to automate this.