Overhaul `client` module
The idea of `Client` was nice, but it had a number of issues, e.g., we
still required users to explicitly call `receive_event` to prevent
buffers from overflowing. It also had bugs, e.g., creating a `Client`
for a non-existing endpoint raised no error, but instead hung.

This patch replaces this with a simpler architecture where users
implement a trait which we can call when something is received from
Zeek.