Allow controlling the size of client outbox and increase default to 256
For ping/pong style events with immediate response the previous default
size of 1 caused the service loop to ping/pong between sending and
receiving which roughly halved the throughput compared to an unbounded
outbox.

This increases the outbox size so both ping/pong as well as send-heavy
clients have reasonable throughput.